requirements.txt
flask
pyserial